requirements.txt
wheel==0.35.1
selenium==3.141.0
config-with-yaml==0.1.0
transliterate==1.10.2
progress==1.5
google-cloud-storage==1.35.0
redis==3.5.3